clear memoized x509 cert on Cert#cert_pem assignment
When a Puma::Acme::Cert is renewed, the new cert is assigned by calling
the cert_pem= method. The instance memoizes the parsed OpenSSL x509
object, but was not taking changes to the cert_pem into account. This
was causing renewed certs to keep using the old expir(ed/ing) cert
instead of the replacement. The pstore library is capable of marshaling
& unmarshaling the @x509 member, so this inconsistency is present in
instances loaded from the pstore cache.